+++ CHANGELOG.md
@@ -9,9 +9,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Changed
 * `kwarray.embed_slice` now does error checking
 * Add `kwarray.RunningStats.update_many` to update with multiple observations at once.
+* Added missing `ArrayAPI.min`, `ArrayAPI.min_argmin` funcs.
 
 ### Fixed
 * Fix ArrayAPI.max for torch in the case where not both items are ints
+* Fixed various ArrayAPI cases, added tests.
 
 
 ## Version 0.6.3 - Released 2022-07-31